requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
configparser==3.5.0
coreapi==2.3.3
coreschema==0.0.4
django-formtools==2.1
Flask==0.12.2
idna==2.7
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.9.6
jsmin==2.2.2
MarkupSafe==1.0
oauthlib==2.1.0
openapi-codec==1.3.2
python-openid==2.2.5
pytz==2018.4
requests==2.19.1
requests-oauthlib==1.0.0
simplejson==3.15.0
uritemplate==3.0.0
urllib3==1.23
virtualenv==15.1.0
Werkzeug==0.13
